requirements.txt
Flask==3.1.2
google-generativeai==0.8.6
python-dotenv==1.2.1
requests==2.32.5